Give SatelliteNotExistError a default message
Don't need to create a new helper method when we can just `raise` it.

+++ lib/gitlab/satellite/satellite.rb
@@ -1,5 +1,9 @@
 module Gitlab
-  class SatelliteNotExistError < StandardError;  end
+  class SatelliteNotExistError < StandardError
+    def initialize(msg = "Satellite doesn't exist")
+      super
+    end
+  end
 
   module Satellite
     class Satellite
@@ -17,12 +21,8 @@ module Gitlab
         Gitlab::Satellite::Logger.error(message)
       end
 
-      def raise_no_satellite
-        raise SatelliteNotExistError.new("Satellite doesn't exist")
-      end
-
       def clear_and_update!
-        raise_no_satellite unless exists?
+        raise SatelliteNotExistError unless exists?
 
         File.exists? path
         @repo = nil
@@ -55,7 +55,7 @@ module Gitlab
       # * Changes the current directory to the satellite's working dir
       # * Yields
       def lock
-        raise_no_satellite unless exists?
+        raise SatelliteNotExistError unless exists?
 
         File.open(lock_file, "w+") do |f|
           begin
@@ -77,7 +77,7 @@ module Gitlab
       end
 
       def repo
-        raise_no_satellite unless exists?
+        raise SatelliteNotExistError unless exists?
 
         @repo ||= Grit::Repo.new(path)
       end
